requirements.txt
apache-airflow-providers-google
pandas
requests
google-cloud-storage
google-cloud-bigquery